requirements.txt
h5py
typing-extensions
wheel
opencv-python
Pillow
scikit-learn
numpy
tqdm
matplotlib
notebook
keras-facenet
mtcnn

#tensorflow==2.13.0
#have to install pytorch before
#pip install openmim
#pip install yolov5 
#mim install mmcv
#python -m pip install --upgrade pip
